+++ krokiet/ui/popup_select_results.slint
@@ -21,7 +21,6 @@ export component PopupSelectResults inherits Rectangle {
     out property <length> all_items_height: item_height * model.length;
 
     popup_window := PopupWindow {
-
         width: item_width;
         height: all_items_height;
